Reorganize tests and fix line breaks in some files
