+++ src/tests/schema/sitei18n.test.ts
@@ -1,5 +1,5 @@
 import { SiteI18nSchema } from '../../schema/sitei18n.js';
-import { describeObjectSchema } from '../_utils.js';
+import { describeObjectSchema } from '../utils.js';
 
 describeObjectSchema(
 	SiteI18nSchema,
